--- a/Code/GraphMol/Bond.cpp
+++ b/Code/GraphMol/Bond.cpp
@@ -177,60 +177,6 @@ double Bond::getBondTypeAsDouble() const {
   }
 }
 
-uint8_t Bond::getTwiceBondType() const {
-  switch (getBondType()) {
-    case UNSPECIFIED:
-    case IONIC:
-    case ZERO:
-      return 0;
-      break;
-    case SINGLE:
-      return 2;
-      break;
-    case DOUBLE:
-      return 4;
-      break;
-    case TRIPLE:
-      return 6;
-      break;
-    case QUADRUPLE:
-      return 8;
-      break;
-    case QUINTUPLE:
-      return 10;
-      break;
-    case HEXTUPLE:
-      return 12;
-      break;
-    case ONEANDAHALF:
-      return 3;
-      break;
-    case TWOANDAHALF:
-      return 5;
-      break;
-    case THREEANDAHALF:
-      return 7;
-      break;
-    case FOURANDAHALF:
-      return 9;
-      break;
-    case FIVEANDAHALF:
-      return 11;
-      break;
-    case AROMATIC:
-      return 3;
-      break;
-    case DATIVEONE:
-      return 2;
-      break;  // FIX: this should probably be different
-    case DATIVE:
-      return 2;
-      break;  // FIX: again probably wrong
-    default:
-      UNDER_CONSTRUCTION("Bad bond type");
-  }
-}
-
 double Bond::getValenceContrib(const Atom *atom) const {
   switch (getBondType()) {
     case UNSPECIFIED:
@@ -353,6 +299,60 @@ void Bond::setStereoAtoms(unsigned int bgnIdx, unsigned int endIdx) {
   atoms.push_back(bgnIdx);
   atoms.push_back(endIdx);
 };
+
+uint8_t getTwiceBondType(const Bond &b) {
+  switch (b.getBondType()) {
+    case Bond::UNSPECIFIED:
+    case Bond::IONIC:
+    case Bond::ZERO:
+      return 0;
+      break;
+    case Bond::SINGLE:
+      return 2;
+      break;
+    case Bond::DOUBLE:
+      return 4;
+      break;
+    case Bond::TRIPLE:
+      return 6;
+      break;
+    case Bond::QUADRUPLE:
+      return 8;
+      break;
+    case Bond::QUINTUPLE:
+      return 10;
+      break;
+    case Bond::HEXTUPLE:
+      return 12;
+      break;
+    case Bond::ONEANDAHALF:
+      return 3;
+      break;
+    case Bond::TWOANDAHALF:
+      return 5;
+      break;
+    case Bond::THREEANDAHALF:
+      return 7;
+      break;
+    case Bond::FOURANDAHALF:
+      return 9;
+      break;
+    case Bond::FIVEANDAHALF:
+      return 11;
+      break;
+    case Bond::AROMATIC:
+      return 3;
+      break;
+    case Bond::DATIVEONE:
+      return 2;
+      break;  // FIX: this should probably be different
+    case Bond::DATIVE:
+      return 2;
+      break;  // FIX: again probably wrong
+    default:
+      UNDER_CONSTRUCTION("Bad bond type");
+  }
+}
 
 };  // namespace RDKit