#### What does this implement/fix? Explain your changes.

The changes in this patch increase throughput of `iterateCIPRanks` by about 62%.  This patch consists of many changes that all combine to contribute to the performance gains, and these are the major changes:

* Introduce `Bond.getTwiceBondType` to avoid floating point operations in `floor(2*bond->getBondTypeAsDouble() + .1)`
* In RDGeneral/Ranking.h/rankVect, just store the last index (instead of the last vector) and avoid creating a copy at the current index.
* In GraphMol/Chirality.cpp/iterateCIPRanks, reserve more capacity ahead of time.
* In GraphMol/Chirality.cpp/iterateCIPRanks, avoid materializing `localEntry` and then copying into `cipEntries`; instead, track the data to be added in a `counts` vector and a `updatedNbrIdxs` vector and directly insert into `cipEntries`.

In `GraphMol/Chirality.cpp/iterateCIPRanks`, we also make several micro-optimizations that were individually measured to give a small performance improvement.

#### Any other comments?

After this change, `iterateCIPRanks` still isn't as fast as we would like, and we are considering a few more changes:
* Introducing https://chromium.googlesource.com/chromium/chromium/+/master/base/stack_container.h to use stack-based vectors (as opposed to your typical heap-based vectors) for faster performance
* Updating iterateCIPRanks to skip re-processing atoms with unique ranks (currently, all atoms are continuously reprocessed even if they already have a unique rank).

Do those proposed follow-on changes sound reasonable?  Do you see alternative ways to improve performance?