@jinpan 2020-10-12
> It would be interesting to know what your test run was calculating, or to see some profiling data.

My benchmark is a retrosynthesis calculation for `O=C1C[C@@H](c2cccc(Cl)c2)C(=O)N1c1nncn1C1CC1`, where we compute `MolToSmiles`/`SmilesToMol` about 1 million times over drug-sized molecules, and I see from the profiler that this patch reduces the amount of cumulative cpu time spent in `iterateCIPRanks` from 68.8s to 42.5s.  This translates to a cumulative cpu time reduction in `MolToSmiles` from 330.6s to 299.4s and `SmilesToMol` from 180.4s to 166.7s.

Is this the level of detail that you'd like to see in the initial PR description?