@greglandrum 2020-11-26
I'm finally starting to look at this now; sorry that took so long.
Before I looked at the code, I tried to get my own sense of the performance improvements and I see roughly the same behavior you do.

My setup is simple and should be biased to make anything that impacts the chirality perception code look good: parse ~430K ChEMBL compounds that have at least one chiral center or stereobond and then generate SMILES for the resulting molecules. I do this in blocks of 10K molecules so that I don't start pushing the amount of memory being used. The script and data are attached below.

Runtime with the current RDKit master is 118.2s to parse the SMILES and 39.0s to generate new SMILES.
Runtime with this PR applied to master is 109.1s to parse the SMILES and 38.3s to generate new SMILES.

So that's a ~8% reduction in runtime (compared to your 9%) for SMILES parsing. I don't see an appreciable change in generating new SMILES, but that's not a surprise since stereochemistry perception doesn't need to be repeated when generating SMILES for molecules that came from the standard parsers.