@jinpan 2020-12-01
Thanks for taking a look!  I adopted your suggestions and moved `getTwiceBondType` to a standalone function in `Bond.h` for other cases that may benefit, such as https://github.com/rdkit/rdkit/blob/master/Code/GraphMol/new_canon.cpp#L415.

I don't see your script/data for benchmarking on this PR - could you try re-attaching?